+++ Assets/Scripts/Projectiles/IProjectile.cs
@@ -4,7 +4,7 @@ namespace AirTowerDefence.Projectile
 {
     internal interface IProjectile
     {
-        void Initialize(Transform target, float damage, float speed);
+        void Initialize(Transform target, int damage, float speed);
     }
 }
